atualizações testes de integração, só falta os de publications

+++ src/medias/medias.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, ConflictException, Injectable, NotFoundException } from '@nestjs/common';
 import { CreateMediaDto } from './dto/create-media.dto';
 import { UpdateMediaDto } from './dto/update-media.dto';
 import { MediasRepository } from './medias.repository';
@@ -7,23 +7,46 @@ import { MediasRepository } from './medias.repository';
 export class MediasService {
   constructor(private readonly mediasRepository: MediasRepository){}
   
-  create(body: CreateMediaDto) {
-    return this.mediasRepository.createMedia(body)
+  async create(body: CreateMediaDto) {
+    if(!body.title || !body.username){
+      throw new BadRequestException();
+    }
+    const existingMedia = await this.mediasRepository.findUsernameAndTitle(body.username, body.title);
+    if(existingMedia !== null){
+      throw new ConflictException();
+    }
+    return await this.mediasRepository.createMedia(body)
   }
 
   findAll() {
     return this.mediasRepository.findAll();
   }
 
-  findOne(id: number) {
-    return this.mediasRepository.findById(id);
+  async findOne(id: number) {
+    const findMedia = await this.mediasRepository.findById(id);
+    if(findMedia === null){
+      throw new NotFoundException();
+    }
+    return findMedia;
   }
 
-  update(id: number, body: UpdateMediaDto) {
+  async update(id: number, body: UpdateMediaDto) {
+    const existingMedia = await this.mediasRepository.findById(id);
+    if(existingMedia === null){
+      throw new NotFoundException();
+    }
+    const existingMediaByTitle = await this.mediasRepository.findUsernameAndTitle(body.username, body.title);
+    if(existingMediaByTitle !== null){
+      throw new ConflictException();
+    }
     return this.mediasRepository.updatedMedia(id, body);
   }
 
-  remove(id: number) {
+  async remove(id: number) {
+    const existingMedia = await this.mediasRepository.findById(id);
+    if(existingMedia === null){
+      throw new NotFoundException();
+    }
     return this.mediasRepository.removeMedia(id);
   }
 }
