+++ src/posts/posts.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
 import { CreatePostDto } from './dto/create-post.dto';
 import { UpdatePostDto } from './dto/update-post.dto';
 import { PostsRepository } from './posts.repository';
@@ -8,22 +8,38 @@ export class PostsService {
   constructor(private readonly postsRepository: PostsRepository){}
   
   create(body: CreatePostDto) {
-    return this.postsRepository.createPost(body);;
+    if(!body.title || !body.text){
+      throw new BadRequestException();
+    }
+    return this.postsRepository.createPost(body);
   }
 
   findAll() {
     return this.postsRepository.findAll();
   }
 
-  findOne(id: number) {
-    return this.postsRepository.findById(id);
+  async findOne(id: number) {
+    const findPost = await this.postsRepository.findById(id);
+    if(findPost === null){
+      throw new NotFoundException();
+    }
+    return findPost;
   }
 
-  update(id: number, body: UpdatePostDto) {
+  async update(id: number, body: UpdatePostDto) {
+    const existingPost = await this.postsRepository.findById(id);
+    if(existingPost === null){
+      throw new NotFoundException();
+    }
+
     return this.postsRepository.updatedPost(id, body);
   }
 
-  remove(id: number) {
+  async remove(id: number) {
+    const existingPost = await this.postsRepository.findById(id);
+    if(existingPost === null){
+      throw new NotFoundException();
+    }
     return this.postsRepository.deletePost(id);
   }
 }
