+++ src/publications/publications.service.ts
@@ -1,29 +1,59 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
 import { CreatePublicationDto } from './dto/create-publication.dto';
 import { UpdatePublicationDto } from './dto/update-publication.dto';
 import { PublicationsRepository } from './publications.repository';
+import { MediasRepository } from '../medias/medias.repository';
+import { PostsRepository } from '../posts/posts.repository';
 
 @Injectable()
 export class PublicationsService {
   constructor(private readonly publicationsRepository: PublicationsRepository){}
   
-  create(body: CreatePublicationDto) {
-    // return this.publicationsRepository.createPubli(body);
+  async create(body: CreatePublicationDto) {
+    if(!body.mediasId || !body.postId || !body.date){
+      console.log('entrou no erro')
+      throw new BadRequestException();
+    }
+    const existingMediaId = await this.publicationsRepository.findExistingMedia(body.mediasId);
+    const existingPostId = await this.publicationsRepository.findExistingPost(body.postId);
+    if(existingMediaId === null || existingPostId === null){
+      throw new NotFoundException();
+    }
+     const result = await this.publicationsRepository.createPubli(body);
+     return result;
   }
 
   findAll() {
     return this.publicationsRepository.findAll();
   }
 
-  findOne(id: number) {
-    return this.publicationsRepository.findById(id);
+  async findOne(id: number) {
+    const findPubli = await this.publicationsRepository.findById(id);
+    if(findPubli === null){
+      throw new NotFoundException();
+    }
+    return findPubli;
   }
 
-  update(id: number, body: UpdatePublicationDto) {
+  async update(id: number, body: UpdatePublicationDto) {
+
+    const findPubli = await this.publicationsRepository.findById(id);
+    const existsMediaId = await this.publicationsRepository.findExistingMedia(body.mediasId);
+    const existisPostId = await this.publicationsRepository.findExistingPost(body.postId);
+
+    if(findPubli === null || existsMediaId === null || existisPostId === null){
+      throw new NotFoundException();
+    }
+    
     return this.publicationsRepository.updatedPubli(id, body);
   }
 
-  remove(id: number) {
+  async remove(id: number) {
+
+    const existingPubli = await this.publicationsRepository.findById(id);
+    if(existingPubli === null){
+      throw new NotFoundException();
+    }
     return this.publicationsRepository.removePubli(id);
   }
 }
